Added base files for the project.
TCPlib is an implementation from a previous project of mine.
It may need to be edited to use protocol buffers.

+++ lab8/v1/server.c
@@ -27,11 +27,11 @@ int main(){
 	
 	arg_1 = 1;
 	pthread_create(&thread_id_1, NULL, thread_code, &arg_1);
-	printf("New thread %d\n", (int) thread_id_1);
+	printf("New thread %lu\n", thread_id_1);
 
 	arg_2 = 2;	
 	pthread_create(&thread_id_2, NULL, thread_code, &arg_2);
-	printf("New thread %d\n", (int) thread_id_2);
+	printf("New thread %lu\n", thread_id_2);
 
 	getchar();
